requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2022.6.15
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.0      
discord-py-slash-command==4.2.1
discord.py==2.0.0
dnspython==2.2.1
easy-pil==0.1.9
ffmpeg==1.4
frozenlist==1.3.0
idna==3.3
motor==3.0.0
multidict==6.0.2
Pillow==8.4.0
pycparser==2.21
pymongo==4.1.1
PyNaCl==1.5.0
python-dotenv==0.20.0
PyYAML==6.0
requests==2.28.1
typing_extensions==4.3.0
urllib3==1.26.12
yarl==1.7.2
youtube-dl==2021.12.17